split expand logic to separate module
